+++ src/main/java/io/github/cpjust/testng_annotations/annotations/EmptySource.java
@@ -1,5 +1,6 @@
 package io.github.cpjust.testng_annotations.annotations;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -23,6 +24,7 @@ import java.lang.annotation.Target;
  *   </li>
  * </ul>
  */
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface EmptySource {