+++ src/main/java/io/github/cpjust/testng_annotations/annotations/NullAndEmptySource.java
@@ -1,5 +1,6 @@
 package io.github.cpjust.testng_annotations.annotations;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -11,6 +12,7 @@ import java.lang.annotation.Target;
  * <p>
  * This is equivalent to using both @NullSource and @EmptySource on the same method.
  */
+@Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface NullAndEmptySource {