+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
-## [1.2.0] - 2025-11-28
+## [1.3.0] - 2026-01-31
+### Added
+- `@EnumSource` annotation to provide enum constants as parameters to test methods, similar to JUnit 5's `@EnumSource`.
+- `EnumSourceListener` to handle the `@EnumSource` annotation and provide test values for parameterized tests.
 
+## [1.2.0] - 2025-12-30
 ### Added
 - `@CsvSource` annotation to provide a source for test values from CSV strings, similar to JUnit 5's `@CsvSource`.
 - `CsvSourceListener` to handle the `@CsvSource` annotation and provide test values for parameterized tests.